+++ docs/changelog.md
@@ -14,9 +14,8 @@ Security in case of vulnerabilities. -->
 
 - [ ] **Add** the ontology normaliser at `deeponto.onto.normaliser`.
 - [ ] **Add** the BERTSubs module at `deeponto.onto.subs`.
-- [ ] **Add** the `subclass_dict` method for `deeponto.onto.Ontology`.
-- [ ] **Add** the `owl_complex_classes` attribute for `deeponto.onto.Ontology`.
 - [ ] **Add** the [detailed instructions](../verbaliser) for how to use the ontology verbaliser. 
+- [X] **Add** the method `remove_axiom` for removing an axiom from the ontology at `deeponto.onto.Ontology` (note that the counterpart `add_axiom` has already been available).
 - [X] **Add** the method `check_named_entity` for checking if an entity is named at `deeponto.onto.Ontology`.
 - [X] **Add** the method `get_subsumption_axioms` for getting subsumption axioms subject to different entity types at `deeponto.onto.Ontology`.
 - [X] **Add** the method `get_asserted_complex_classes` for getting all complex classes that occur in ontology (subsumption and/or equivalence) axioms at `deeponto.onto.Ontology`.